Fixing selection for zooms!=100%.
svn path=/branches/kpdf_experiments/kdegraphics/kpdf/; revision=359121

+++ kpdf/pageviewutils.h
@@ -35,7 +35,7 @@ class PageViewItem
             m_geometry.setHeight( h );
             m_zoomFactor = z;
         }
-        float zoomFactor() { return m_zoomFactor; }
+        double zoomFactor() { return m_zoomFactor; }
 
         int pageNumber() const { return m_page->number(); }
         const KPDFPage * page() const { return m_page; }
